+++ drone/src/wall_time.cpp
@@ -3,13 +3,13 @@
 WallTime wall_time;
 static auto _t_start = std::chrono::steady_clock::now();
 
-double WallTime::millis()
+uint32_t WallTime::millis()
 {
     auto t_now = std::chrono::steady_clock::now();
     return std::chrono::duration<double, std::milli>(t_now - _t_start).count();
 }
 
-double WallTime::micros()
+uint32_t WallTime::micros()
 {
     auto t_now = std::chrono::steady_clock::now();
     return std::chrono::duration<double, std::micro>(t_now - _t_start).count();
